Refactor TeXAtom, mglyph, and mspace into common and chtml components
